refactor(repository): reorganize relations source code into smaller files
This commit does the following:
- move the definition of relation metadata into a standalone file
- move relation decorators into standalone files
- move relation repositories into standalone files
- move relation factories into standalone files
