+++ packages/repository/src/relations/has-many/has-many.repository.ts
@@ -3,20 +3,16 @@
 // This file is licensed under the MIT License.
 // License text available at https://opensource.org/licenses/MIT
 
-import {Count, DataObject, Options} from '../common-types';
-import {Entity} from '../model';
-import {Filter, Where} from '../query';
+import {Getter} from '@loopback/context';
+import {Count, DataObject, Options} from '../../common-types';
+import {Entity} from '../../model';
+import {Filter, Where} from '../../query';
 import {
   constrainDataObject,
   constrainFilter,
   constrainWhere,
-} from './constraint-utils';
-import {EntityCrudRepository} from './repository';
-import {Getter} from '@loopback/context';
-import {EntityNotFoundError} from '../errors';
-
-// Re-export Getter so that users don't have to import from @loopback/context
-export {Getter};
+} from '../../repositories/constraint-utils';
+import {EntityCrudRepository} from '../../repositories/repository';
 
 /**
  * CRUD operations for a target repository of a HasMany relation
@@ -58,17 +54,6 @@ export interface HasManyRepository<Target extends Entity> {
     where?: Where<Target>,
     options?: Options,
   ): Promise<Count>;
-}
-
-/**
- * CRUD operations for a target repository of a BelongsTo relation
- */
-export interface BelongsToRepository<Target extends Entity> {
-  /**
-   * Gets the target model instance
-   * @param options
-   */
-  get(options?: Options): Promise<Target>;
 }
 
 export class DefaultHasManyRepository<
@@ -128,36 +113,5 @@ export class DefaultHasManyRepository<
       constrainWhere(where, this.constraint as Where<TargetEntity>),
       options,
     );
-  }
-}
-
-export class DefaultBelongsToRepository<
-  TargetEntity extends Entity,
-  TargetId,
-  TargetRepository extends EntityCrudRepository<TargetEntity, TargetId>
-> implements BelongsToRepository<TargetEntity> {
-  /**
-   * Constructor of DefaultBelongsToEntityCrudRepository
-   * @param getTargetRepository the getter of the related target model repository instance
-   * @param constraint the key value pair representing foreign key name to constrain
-   * the target repository instance
-   */
-  constructor(
-    public getTargetRepository: Getter<TargetRepository>,
-    public constraint: DataObject<TargetEntity>,
-  ) {}
-
-  async get(options?: Options): Promise<TargetEntity> {
-    const targetRepo = await this.getTargetRepository();
-    const result = await targetRepo.find(
-      constrainFilter(undefined, this.constraint),
-      options,
-    );
-    if (!result.length) {
-      // We don't have a direct access to the foreign key value here :(
-      const id = 'constraint ' + JSON.stringify(this.constraint);
-      throw new EntityNotFoundError(targetRepo.entityClass, id);
-    }
-    return result[0];
   }
 }
