+++ packages/repository/src/relations/has-many/index.ts
@@ -1,0 +1,8 @@
+// Copyright IBM Corp. 2018. All Rights Reserved.
+// Node module: @loopback/repository
+// This file is licensed under the MIT License.
+// License text available at https://opensource.org/licenses/MIT
+
+export * from './has-many.decorator';
+export * from './has-many.repository';
+export * from './has-many-repository.factory';
