+++ packages/repository/src/relations/has-one/has-one.decorator.ts
@@ -1,0 +1,17 @@
+// Copyright IBM Corp. 2018. All Rights Reserved.
+// Node module: @loopback/repository
+// This file is licensed under the MIT License.
+// License text available at https://opensource.org/licenses/MIT
+
+import {relation} from '../relation.decorator';
+import {RelationType} from '../relation.types';
+
+/**
+ * Decorator for hasOne
+ * @param definition
+ * @returns {(target:any, key:string)}
+ */
+export function hasOne(definition?: Object) {
+  const rel = Object.assign({type: RelationType.hasOne}, definition);
+  return relation(rel);
+}
