+++ packages/repository/src/relations/index.ts
@@ -1,0 +1,10 @@
+// Copyright IBM Corp. 2017. All Rights Reserved.
+// Node module: @loopback/repository
+// This file is licensed under the MIT License.
+// License text available at https://opensource.org/licenses/MIT
+
+export * from './relation.types';
+export * from './relation.decorator';
+export * from './belongs-to';
+export * from './has-many';
+export * from './has-one';
