+++ packages/repository/src/relations/relation.decorator.ts
@@ -1,0 +1,77 @@
+// Copyright IBM Corp. 2018. All Rights Reserved.
+// Node module: @loopback/repository
+// This file is licensed under the MIT License.
+// License text available at https://opensource.org/licenses/MIT
+
+import {PropertyDecoratorFactory} from '@loopback/context';
+import {Model, RelationDefinitionMap} from '../model';
+import {RelationType} from './relation.types';
+
+export const RELATIONS_KEY = 'loopback:relations';
+
+/**
+ * Decorator for relations
+ * @param definition
+ * @returns {(target:any, key:string)}
+ */
+export function relation(definition?: Object) {
+  // Apply relation definition to the model class
+  return PropertyDecoratorFactory.createDecorator(RELATIONS_KEY, definition);
+}
+
+/**
+ * Get metadata of all relations defined on a given model class.
+ *
+ * @param modelCtor The model class (the constructor function).
+ * @return
+ */
+export function getModelRelations(
+  modelCtor: typeof Model,
+): RelationDefinitionMap {
+  return (modelCtor.definition && modelCtor.definition.relations) || {};
+}
+
+//
+// placeholder decorators for relations that are not implemented yet
+// TODO: move these decorators to per-relation subdirectories
+//
+
+/**
+ * Decorator for embedsOne
+ * @param definition
+ * @returns {(target:any, key:string)}
+ */
+export function embedsOne(definition?: Object) {
+  const rel = Object.assign({type: RelationType.embedsOne}, definition);
+  return PropertyDecoratorFactory.createDecorator(RELATIONS_KEY, rel);
+}
+
+/**
+ * Decorator for embedsMany
+ * @param definition
+ * @returns {(target:any, key:string)}
+ */
+export function embedsMany(definition?: Object) {
+  const rel = Object.assign({type: RelationType.embedsMany}, definition);
+  return PropertyDecoratorFactory.createDecorator(RELATIONS_KEY, rel);
+}
+
+/**
+ * Decorator for referencesOne
+ * @param definition
+ * @returns {(target:any, key:string)}
+ */
+export function referencesOne(definition?: Object) {
+  const rel = Object.assign({type: RelationType.referencesOne}, definition);
+  return PropertyDecoratorFactory.createDecorator(RELATIONS_KEY, rel);
+}
+
+/**
+ * Decorator for referencesMany
+ * @param definition
+ * @returns {(target:any, key:string)}
+ */
+export function referencesMany(definition?: Object) {
+  const rel = Object.assign({type: RelationType.referencesMany}, definition);
+  return PropertyDecoratorFactory.createDecorator(RELATIONS_KEY, rel);
+}
