+++ packages/repository/src/repositories/index.ts
@@ -7,6 +7,4 @@ export * from './kv.repository';
 export * from './legacy-juggler-bridge';
 export * from './kv.repository.bridge';
 export * from './repository';
-export * from './relation.factory';
-export * from './relation.repository';
 export * from './constraint-utils';
